refactor loading weights from remote instance coding format
